requirements.txt
flet==0.22.*
pillow==9.5.0